+++ src/create-context.ts
@@ -1,5 +1,17 @@
 /* eslint-disable prefer-object-spread */
-import { Context } from 'unleash-client';
+import type { Context } from 'unleash-client';
+
+function tryParseDate(dateString: string | undefined): Date | undefined {
+    if (!dateString) {
+        return undefined;
+    }
+    const parsedDate = new Date(dateString);
+    if (!Number.isNaN(parsedDate.getTime())) {
+        return parsedDate;
+    } else {
+        return undefined;
+    }
+}
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export function createContext(value: any): Context {
@@ -10,6 +22,7 @@ export function createContext(value: any): Context {
         sessionId,
         remoteAddress,
         properties,
+        currentTime,
         ...rest
     } = value;
 
@@ -20,6 +33,7 @@ export function createContext(value: any): Context {
         userId,
         sessionId,
         remoteAddress,
+        currentTime: tryParseDate(currentTime),
         properties: Object.assign({}, rest, properties),
     };