+++ src/test/create-context.test.ts
@@ -25,6 +25,29 @@ test('should move rest props to properties', () => {
     expect(context.properties?.tenantId).toBe('some-tenant');
 });
 
+test('when an extra property is both on the top-level and under properties, the properties-level wins', () => {
+    const context = createContext({
+        customProperty: 'top-level',
+        properties: {
+            customProperty: 'properties-level',
+        },
+    });
+
+    expect(context).not.toHaveProperty('customProperty');
+    expect(context.properties?.customProperty).toBe('properties-level');
+});
+
+test('If you specify top-level properties under properties, they do not get moved up', () => {
+    const context = createContext({
+        properties: {
+            appName: 'name',
+        },
+    });
+
+    expect(context.properties?.appName).toBe('name');
+    expect(context.appName).toBe(undefined);
+});
+
 test('should keep properties', () => {
     const context = createContext({
         userId: '123',
@@ -58,6 +81,31 @@ test('will not blow up if properties is an array', () => {
     expect(context.userId).toBe('123');
     expect(context).not.toHaveProperty('tenantId');
     expect(context).not.toHaveProperty('region');
+});
+
+test('accepts current time as a context value', () => {
+    const targetDate = new Date('2024-01-01T00:00:00.000Z');
+    const context = createContext({
+        currentTime: targetDate.toISOString(),
+    });
+
+    expect(context.currentTime).toStrictEqual(targetDate);
+});
+
+test('invalid time strings fall back to undefined currentTime', () => {
+    const context = createContext({
+        currentTime: 'its cute that you think this will parse',
+    });
+
+    expect(context).not.toHaveProperty('currentTime');
+});
+
+test('missing time current time falls back to undefined currentTime', () => {
+    const context = createContext({
+        userId: '123',
+    });
+
+    expect(context).not.toHaveProperty('currentTime');
 });
 
 test.skip('will not blow up if userId is an array', () => {