+++ CHANGELOG.md
@@ -1,5 +1,38 @@
 # Changelog
 
+### 0.8.0
+- feat: loading of custom strategies when using Docker (#56)
+- feat: upgrade node-sdk to v3.12.0 (#57)
+- chore: Create LICENSE (#58)
+### 0.7.1
+- fix: add start on metrics
+
+### 0.7.0
+
+- feat: generate type output to publish them to npm (#54)
+- feat: experimental support for bootstrap (#44)
+- fix: upgrade unleash-client to v3.11.3
+- fix: require authorization header for /client/metrics (#55)
+- fix: do not crash for invalid trustProxy option (#50)
+
+### 0.6.1
+- fix: upgrade unleash-client to v3.11.2
+### 0.6.0
+- feat: propagate impression data field (#48)
+- fix: remove npm folder in docker image after build
+- fix: export Client and createProxyConfig (#52)
+- fix: Create codeql-analysis.yml
+- docs: Fix description of `proxy_base_path` to match implementation (#51)
+- docs: fix env variable names so the docs match the implementation. (#43)
+
+### 0.5.0
+- feat: additional possibility to customize header for proxy secrets (#39)
+- fix: POST should handle empty toggle names
+- fix: pin dependencies
+- fix: Build multi-arch Docker container to support arm64 and amd64 (#30)
+- docs: fix typo in `docker pull` command (#26)
+- docs: Add proxyPort/PORT to the list of options/env vars (#28)
+
 ### 0.4.0
 - feat: add compression #23